requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
requests==2.32.3
beautifulsoup4==4.12.3
pytest==8.3.5
